requirements.txt
streamlit==1.32.0
streamlit-authenticator==0.3.1
supabase==2.4.1
python-dotenv==1.0.0
pandas==2.2.0
plotly==5.18.0
Pillow==10.2.0
openpyxl==3.1.2
reportlab==4.1.0
python-multipart==0.0.9
